Generated noops.yaml with Path() support
Add Path() scalar representer as much as human readable possible.
A Path() will be represented in a yaml file with:

!path '/a/path/to/something'

Add associated constuctor to load !path as a Path()

JSON is excluded as it is mainly used for external script with jq so
json output will always represent Path() as str.